backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.26.2
pyautogui==0.9.54
screen-brightness-control==0.22.0
pycaw==20230407
comtypes==1.2.0
python-dotenv==1.0.0
supabase==2.3.0
gspread==5.12.0
oauth2client==4.1.3
